--- src/test_helpers/contracts.rs
@@ -1,9 +1,0 @@
-//! Module for simple contracts to be used in tests
-
-pub mod caller;
-pub mod echo;
-pub mod error;
-pub mod hackatom;
-pub mod payout;
-pub mod reflect;
-pub mod stargate;
