+++ src/testing_helpers/caller.rs
@@ -1,11 +1,9 @@
-use std::fmt;
-
+use crate::{Contract, ContractWrapper};
 use cosmwasm_std::{
     Binary, Deps, DepsMut, Empty, Env, MessageInfo, Response, StdError, SubMsg, WasmMsg,
 };
 use schemars::JsonSchema;
-
-use crate::{Contract, ContractWrapper};
+use std::fmt::Debug;
 
 fn instantiate(
     _deps: DepsMut,
@@ -35,7 +33,7 @@ fn query(_deps: Deps, _env: Env, _msg: Empty) -> Result<Binary, StdError> {
 
 pub fn contract<C>() -> Box<dyn Contract<C>>
 where
-    C: Clone + fmt::Debug + PartialEq + JsonSchema + 'static,
+    C: Clone + Debug + PartialEq + JsonSchema + 'static,
 {
     let contract = ContractWrapper::new_with_empty(execute, instantiate, query);
     Box::new(contract)
